refactor(app): useRobotInitializationStatus gets its own file
