+++ app/src/resources/health/useRobotInitializationStatus.ts
